+++ src/pudl/workspace/datastore.py
@@ -190,7 +190,7 @@ class ZenodoDoiSettings(BaseSettings):
     eia176: ZenodoDoi = "10.5281/zenodo.10607843"
     eia191: ZenodoDoi = "10.5281/zenodo.10607837"
     eia757a: ZenodoDoi = "10.5281/zenodo.10607839"
-    eia860: ZenodoDoi = "10.5281/zenodo.10067566"
+    eia860: ZenodoDoi = "10.5281/zenodo.11662381"
     eia860m: ZenodoDoi = "10.5281/zenodo.11110602"
     eia861: ZenodoDoi = "10.5281/zenodo.10204708"
     eia923: ZenodoDoi = "10.5281/zenodo.11111206"